// src/services/gitlab/openapi.d.ts
export interface paths {
// ...
    "/api/v4/projects/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Retrieve a project
         * @description Retrieves details on a specified project. This endpoint can be accessed without authentication if the project is publicly accessible.
         */
        get: operations["getApiV4ProjectsId"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}